agents/video-generator/requirements.txt
google-genai>=1.0.0
supabase>=2.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
modal>=0.73.0
